gh-108303: Move more files to `Lib/test/test_module`
