+++ Lib/test/test_module/final_b.py
@@ -3,7 +3,7 @@ Fodder for module finalization tests in test_module.
 """
 
 import shutil
-import test.final_a
+import test.test_module.final_a
 
 x = 'b'
 
@@ -11,7 +11,7 @@ class C:
     def __del__(self):
         # Inspect module globals and builtins
         print("x =", x)
-        print("final_a.x =", test.final_a.x)
+        print("final_a.x =", test.test_module.final_a.x)
         print("shutil.rmtree =", getattr(shutil.rmtree, '__name__', None))
         print("len =", getattr(len, '__name__', None))
